bugfix
not yet complete, but recognizes if theres more letters on one image. For example, if there is 2 letters, which are over each other (their projection, not the letters itself) then the segmentation cant recognize it easily, so i had to do this in another way. does not recognize letter 'i', '?', '!' as 2 characters.